Fixed displaying appointments on the main page
In case an appointment is made and a patient or a doctor linked
to that appointment is later deleted, the appointment remains in the
database (with the ID of the patient or the doctor). While trying to
display a list of appointments on the main page, an error was occurring
because I was trying to access fields like first name. The proposed
solution fixes the problem by sending to the main page only the list of
appointments with existing patients and doctors (filtering happens in
the AppointmentService class).